requirements.txt
huggingface_hub
streamlit
python-dotenv
pydantic
smolagents
truststore
requests
certifi
